refactor: extract preset editor for reuse across surfaces
